toolkit/components/url-classifier/tests/mochitest/test_classifier_worker.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test the URI Classifier</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>

<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">

<script class="testbody" type="text/javascript">

var Cc = SpecialPowers.Cc;
var Ci = SpecialPowers.Ci;

// Add some URLs to the malware database.
var testData = "example.com/tests/toolkit/components/url-classifier/tests/mochitest/evilWorker.js";
var testUpdate =
  "n:1000\ni:test-malware-simple\nad:550\n" +
  "a:550:32:" + testData.length + "\n" +
  testData;

var dbService = Cc["@mozilla.org/url-classifier/dbservice;1"]
                .getService(Ci.nsIUrlClassifierDBService);

function doUpdate(update) {
  var listener = {
    QueryInterface: function(iid)
    {
      if (iid.equals(Ci.nsISupports) ||
          iid.equals(Ci.nsIUrlClassifierUpdateObserver))
        return this;

      throw Cr.NS_ERROR_NO_INTERFACE;
    },
    updateUrlRequested: function(url) { },
    streamFinished: function(status) { },
    updateError: function(errorCode) {
      ok(false, "Couldn't update classifier.");
      // Abort test.
      SimpleTest.finish();
    },
    updateSuccess: function(requestedTimeout) {
      SpecialPowers.pushPrefEnv(
        {"set" : [["browser.safebrowsing.malware.enabled", true]]},
        function loadTestFrame() {
          document.getElementById("testFrame").src =
            "http://example.com/tests/toolkit/components/url-classifier/tests/mochitest/workerFrame.html";
        }
      );
    }
  };

  dbService.beginUpdate(listener, "test-malware-simple", "");
  dbService.beginStream("", "");
  dbService.updateStream(update);
  dbService.finishStream();
  dbService.finishUpdate();
}

function onmessage(event)
{
  var pieces = event.data.split(':');
  if (pieces[0] == "finish") {
    SimpleTest.finish();
    return;
  }

  is(pieces[0], "success", pieces[1]);
}

SpecialPowers.pushPrefEnv(
  {"set" : [["urlclassifier.malwareTable", "test-malware-simple"],
            ["urlclassifier.phishTable", "test-phish-simple"]]},
  function() { doUpdate(testUpdate); });

window.addEventListener("message", onmessage, false);

SimpleTest.waitForExplicitFinish();

</script>

</pre>
<iframe id="testFrame" onload=""></iframe>
</body>
</html>
